Added license Moved utility methods into a module (no API change)
