+++ .hgignore
@@ -1,0 +1,5 @@
+syntax: glob
+build/*
+dist/*
+*.pyc
+*.egg-info
