+++ setup.py
@@ -1,7 +1,26 @@
+#
+#      Licensed to the Apache Software Foundation (ASF) under one
+#      or more contributor license agreements.  See the NOTICE file
+#      distributed with this work for additional information
+#      regarding copyright ownership.  The ASF licenses this file
+#      to you under the Apache License, Version 2.0 (the
+#      "License"); you may not use this file except in compliance
+#      with the License.  You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#      Unless required by applicable law or agreed to in writing,
+#      software distributed under the License is distributed on an
+#      "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+#      KIND, either express or implied.  See the License for the
+#      specific language governing permissions and limitations
+#      under the License.
+#
+
 import os
 from setuptools import setup, find_packages
 
-version = '0.1'
+version = '0.2'
 
 def read(fname):
     return open(os.path.join(os.path.dirname(__file__), fname)).read()
@@ -11,15 +30,15 @@ setup(
     description = 'Apache Chemistry CMIS client library Alfresco extension for Python',
     version = version,
     author = 'Patrice Collardez',
-    author_email = 'pcollardez@acxio.fr',
+    author_email = 'pcollardez@gmail.com',
     license = 'Apache License (2.0)',
-    url = 'http://www.acxio.fr/',
+    url = 'http://code.google.com/p/cmislib-alfresco-extension/',
     package_dir = {'':'src'},
     packages = find_packages('src', exclude=['tests']),
     #include_package_data = True,
     exclude_package_data = {'':['tests']},
     zip_safe = False,
-    install_requires = ('cmislib'),
+    install_requires = ('cmislib>=0.4'),
     long_description = read('README.txt'),
     classifiers = [
         "Development Status :: 3 - Alpha",
