+++ src/tests/cmislibalftest.py
@@ -1,8 +1,22 @@
-'''
-Created on 27 mai 2011
+#
+#      Licensed to the Apache Software Foundation (ASF) under one
+#      or more contributor license agreements.  See the NOTICE file
+#      distributed with this work for additional information
+#      regarding copyright ownership.  The ASF licenses this file
+#      to you under the Apache License, Version 2.0 (the
+#      "License"); you may not use this file except in compliance
+#      with the License.  You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#      Unless required by applicable law or agreed to in writing,
+#      software distributed under the License is distributed on an
+#      "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+#      KIND, either express or implied.  See the License for the
+#      specific language governing permissions and limitations
+#      under the License.
+#
 
-@author: pcollardez
-'''
 import unittest
 from unittest import TestSuite, TestLoader
 import cmislibalf
@@ -46,7 +60,7 @@ class HookTest(CmisAlfTestBase):
         titledAspectObjectType = self._repo.getTypeDefinition('P:sys:temporary')
         self.assertFalse(otherDoc.hasAspect(titledAspectObjectType))
         
-        self.assertEqual(3, len(otherDoc.getAspects()))
+        self.assertEqual(4, len(otherDoc.getAspects()))
         
         aspect = otherDoc.findAspect('cm:title')
         self.assertEqual('P:cm:titled', aspect.getTypeId())
